# src/main/java/br/com/estoque/gerenciamento/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "REST API Gerenciamente: Produtos e Usuários",
    "description": "# 🚀 Sobre\n\nEsta documentação tem como objetivo explicar o funcionamento de cada operação CRUD que a API Gerenciamentos faz.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "localhost://"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
  "paths": {
...
    "/191/produtos/{id}": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "summary": "Listar {id}",
        "description": "Encontra um produto especifico na lista de produtos com base no seu id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "put": {
        "tags": [
          "Produtos"
        ],
        "summary": "Atualizar",
        "description": "Encontra um produto específico e o atualiza.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "nome": "Baré",
                  "descricao": "Refrigerante",
                  "preco": 5.5,
                  "codigo": "159",
                  "fornecedor": "Baré",
                  "categoria": "bebidas",
                  "validade": "30/01/2025",
                  "image": "Sem imagem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  }
}